// nautical-data/code/nautical-maps/src/types/types.gen.ts
export type Terminal = Location & {
    geoCoordinate: Coordinate;
    /**
     * Port facility number of the terminal
     */
    portFacilityNumber?: string;
    /**
     * Type of the terminal
     */
    terminalType: 'Unknown' | 'BulkTerminal' | 'ContainerTerminal' | 'TankerTerminal' | 'FerryTerminal' | 'PassengerTerminal' | 'RoRoTerminal' | 'ServiceHarbour' | 'ServiceRepair' | 'Shipyard';
};
// ...
export type Coordinate = {
    /**
     * Latitude part of coordinate
     */
    latitude: number;
    /**
     * Longitude part of coordinate
     */
    longitude: number;
};

/**
 * A terminal with berths
 */
export type TerminalWithBerths = Terminal & {
    berths: ArrayOfLocation;
};

/**
 * A terminal with berth and berth positions
 */
export type TerminalWithBerthsAndPositions = Terminal & {
    berths: ArrayOfBerthWithPositions;
};

/**
 * A berth with berth positions
 */
export type BerthWithPositions = Berth & {
    positions?: ArrayOfBerthPosition;
};

export type ArrayOfCoordinate = Array<Coordinate>;

export type ArrayOfLocation = Array<Location>;

export type ArrayOfBerthWithPositions = Array<BerthWithPositions>;

export type ArrayOfBerthPosition = Array<BerthPosition>;

export type ArrayOfTerminalWithBerthsAndPositions = Array<TerminalWithBerthsAndPositions>;